+++ KikoleSite/Repositories/IReadRepository.cs
@@ -10,6 +10,6 @@ namespace KikoleSite.Repositories
     {
         Task<IReadOnlyCollection<PlayerDto>> GetPlayersAsync(bool banned = false, bool fromCache = false);
         Task<IReadOnlyCollection<EntryDto>> GetEntriesAsync(Stage? stage, Level? level, DateTime? startDate, DateTime? endDate);
-        Task<IReadOnlyCollection<EntryDto>> GetPlayerEntriesAsync(long playerId, Game game);
+        Task<IReadOnlyCollection<EntryDto>> GetPlayerEntriesAsync(uint playerId, Game game);
     }
 }
